refactor(http): split gateway/namespace.rs into constants/encode/bare
Break the 648-line gateway/namespace.rs monolith into a thin facade

backed by four focused siblings:

- namespace_constants.rs: name lists, separators, prefix predicates

- namespace_encode.rs: skill.tool + gateway id8.tool encoder/decoder

- namespace_bare.rs: resolve_bare_names (#307) + one-shot warn helpers

- namespace_tests.rs: unit tests

Behaviour unchanged; the facade re-exports every public symbol so

callers keep using crate::gateway::namespace::* unchanged. 266 lib

tests + doctests pass. Updates docs/guide/architecture.md with the

new maintainer layout.
