+++ docs/guide/architecture.md
@@ -296,6 +296,7 @@ dcc-mcp-server ← dcc-mcp-http
 - `src/prompts.rs` keeps the `PromptRegistry` (lazy cache, skill-set invalidation, `list` / `get` surface); the YAML spec types + `PromptError` live in `prompts_spec.rs`, the `{{name}}` template engine lives in `prompts_template.rs`, the sibling-file / glob loader plus the workflow-derived prompt generator live in `prompts_loader.rs`, and unit tests live in `prompts_tests.rs`.
 - `src/protocol.rs` is a thin facade keeping protocol-version negotiation + session/method constants; every MCP message type is split by primitive: JSON-RPC envelope + error codes in `protocol_jsonrpc.rs`, lifecycle (`initialize` / capabilities / roots / logging / elicitation) in `protocol_lifecycle.rs`, tools (`tools/list` / `tools/call` / annotations / content) in `protocol_tools.rs`, resources (`resources/list|read|subscribe`) in `protocol_resources.rs`, prompts (`prompts/list|get`) in `protocol_prompts.rs`, and SSE formatter + cursor helpers in `protocol_sse.rs`.
 - `src/handler.rs` is a thin facade over the top-level axum handlers: shared `AppState` + cancellation/elicitation TTL constants live in `handler_state.rs`, the three HTTP verbs (`POST` / `GET` / `DELETE /mcp`) live in `handler_routes.rs`, notification and response-message routing (`notifications/cancelled`, `roots/list_changed`, elicitation correlation) live in `handler_notifications.rs`, and the JSON-RPC method router plus `initialize` / `tools/list` implementations live in `handler_dispatch.rs`. Per-method request handlers (`tools/call`, `resources/*`, `prompts/*`, `elicitation/create`, …) continue to live under `src/handlers/` and are re-exported through the facade so existing call sites keep using `crate::handler::*`.
+- `src/gateway/namespace.rs` is a thin facade over the tool-name namespace helpers: the canonical name lists (`GATEWAY_LOCAL_TOOLS`, `CORE_TOOL_NAMES`), SEP-986 separator constants, and the `is_local_tool` / `is_core_tool` / `instance_short` / `is_instance_prefix` predicates live in `namespace_constants.rs`; the encoder / decoder pair (`extract_bare_tool_name`, `skill_tool_name`, `decode_skill_tool_name`, `encode_tool_name`, `decode_tool_name`, `assert_gateway_tool_name`) lives in `namespace_encode.rs`; the #307 bare-name resolver (`BareNameInput`, `resolve_bare_names`) and the one-shot deprecation warn helpers (`warn_legacy_prefixed_once`, process-local dedupe state) live in `namespace_bare.rs`; unit tests live in `namespace_tests.rs`. The facade re-exports every public symbol so downstream modules and tests keep using `crate::gateway::namespace::*` unchanged.
 
 ### dcc-mcp-server
